Add CSS loaders from server

Add/remove CSS loaders from server to any Shiny/HTML component. It is useful to indicate busy status when some code is running in the server and when it finishes, remove the loader to indicate clear status.

Value

CSS load in R6 class

Methods

Public methods

addLoader$new()

addLoader$show()

addLoader$hide()

addLoader$clone()

Method new()

Usage

addLoader$new(
  target_selector = "",
  isID = TRUE,
  type = "default",
  id = "",
  height = NULL,
  width = height,
  color = "#337ab7",
  opacity = 1,
  method = "replace",
  block = TRUE,
  center = TRUE,
  bg_color = "#eee",
  footer = NULL,
  z_index = 2000,
  alert = FALSE,
  session = shiny::getDefaultReactiveDomain()
)

Arguments

target_selector

string, which Shiny component you want to add the loader to? a shiny component ID or a valid CSS selector if isID = FLASE. for example, you have a button and want to add animation to it:
actionButton(inputId = "btn")

This function is used in server only, so if you are in shiny module, DO NOT add the ns() wrapper.

isID

bool, is your selector an ID?

type

string, one of "circle", "dual-ring", "facebook", "heart", "ring", "roller", "default", "ellipsis", "grid", "hourglass", "ripple", "spinner", default is "default".

id

string, the unqiue ID for the loader, if not provided, a random ID will be given. If you are using shiny modules, use session$ns('YOUR_ID') to wrap it.

height

string, (r)em, "1.5rem", "1.5em", or pixel, like "10px". Default is NULL, will be automatically calculated based on the target component. It is recommend to use NULL for "replace" and "inline" method to let it automatically be calculated, but required for "full_screen" method.

width

string, default is the same as height to make it sequare.

color

string, any valid CSS color name, or hex color code

opacity

number, between 0-1

method

one of "replace", "inline", "full_screen", see details

block

bool, for some input components, once the loader starts, it can also block user interaction with the component, very useful for "inline" method, eg. prevent users from clicking the button while some process is still running.

center

bool, try to place the load to the center of the target for "inline" and "replace" and center of the screen for "full_screen".

bg_color

string, any valid CSS color name, or hex color code. Only works for "full_screen" method.

footer

Additional Shiny/HTML component to add below the loader, like a title h1("load title").

z_index

number, only works for "full_screen" method, what CSS layer should the overlay be places. In HTML, all elements have the default of 0.

alert

bool, should alert if target cannot be found or other javascript errors? mainly for debugging

session

shiny session

Method show()

Usage

addLoader$show(alert = FALSE)

Method hide()

Usage

addLoader$hide(alert = FALSE)

Method clone()

The objects of this class are cloneable with this method.

Usage

addLoader$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.