Bootstrap3 alert

Add a Bootstrap3 alert component to the UI

bsAlert(..., status = "success", closeable = TRUE)

Arguments

...

any shiny tag or tagList you want to add to the alert body, or any additional attributes you want to add to the alert element.

status

string, one of "success", "info", "warning", "danger"

closeable

bool, can the alert be closed?

Value

shiny tag element

Details

Read more here: https://getbootstrap.com/docs/3.3/components/#alerts

Examples

if(interactive()) {
  library(shiny)
  ui <- fluidPage(
    bsAlert(tags$b("Success: "), "You made it", status = "success"),
    bsAlert(tags$b("Info: "), "Something happened", status = "info"),
    bsAlert(tags$b("Warning: "), "Something is not right", status = "warning"),
    bsAlert(tags$b("Danger: "), "Oh no...", status = "danger")
  )
  server <- function(input, output, session) {}
  shinyApp(ui, server)
}