Add commonly used HTML dependencies

Mostly used in SPS internal development

spsDepend(dep, js = TRUE, css = TRUE)

Arguments

dep
dependency names, see details

js
bool, use only javascript from this resource if there are both js and css files?

css
bool, use only CSS from this resource if there are both js and css files?

Value

htmltools::htmlDependency object

Details

For dep, current options are:

basic: spsComps basic css and js

shinydashboard: shinydashboard package, css and js

AdminLTE: shinydashboard package, often used together with "shinydashboard", css and js

update_pg: spsComps pgPaneUpdate function required, js only

update_timeline: spsComps spsTimeline function required, js only

font-awesome: font-awesome, css only

bttn: comes from shinyWidgets package, bttn.css, css only

toastr: comes from shinytoastr package, toastr.js, css and js

pop-tip: enable enhanced bootstrap popover and tips, required for bsHoverPopover function

gotop: required by spsGoTop function

Examples

spsDepend("basic")

#> [[1]]
#> List of 10
#>  $ name      : chr "spsComps-js"
#>  $ version   : chr "0.1.0.9"
#>  $ src       :List of 2
#>   ..$ href: chr "spsComps"
#>   ..$ file: chr "assets"
#>  $ meta      : NULL
#>  $ script    : chr "js/sps-comps.js"
#>  $ stylesheet: NULL
#>  $ head      : NULL
#>  $ attachment: NULL
#>  $ package   : chr "spsComps"
#>  $ all_files : logi FALSE
#>  - attr(*, "class")= chr "html_dependency"
#> 
#> [[2]]
#> List of 10
#>  $ name      : chr "spsComps-css"
#>  $ version   : chr "0.1.0.9"
#>  $ src       :List of 2
#>   ..$ href: chr "spsComps"
#>   ..$ file: chr "assets"
#>  $ meta      : NULL
#>  $ script    : NULL
#>  $ stylesheet: chr "css/sps-comps.css"
#>  $ head      : NULL
#>  $ attachment: NULL
#>  $ package   : chr "spsComps"
#>  $ all_files : logi FALSE
#>  - attr(*, "class")= chr "html_dependency"
#> 

# shinydashboard has both js and css and if we only want to use css:
spsDepend("shinydashboard", css = FALSE)

#> [[1]]
#> List of 10
#>  $ name      : chr "shinydashboard-js"
#>  $ version   : chr "0.7.1"
#>  $ src       :List of 1
#>   ..$ file: chr "/home/runner/work/_temp/Library/shinydashboard"
#>  $ meta      : NULL
#>  $ script    : chr "shinydashboard.js"
#>  $ stylesheet: NULL
#>  $ head      : NULL
#>  $ attachment: NULL
#>  $ package   : NULL
#>  $ all_files : logi TRUE
#>  - attr(*, "class")= chr "html_dependency"
#> 
#> [[2]]
#> NULL
#> 

# Then add it to your shiny app
if(interactive()){
    library(shiny)

    ui <- fluidPage(
      tags$i(class = "fa fa-home"),
      spsDepend("font-awesome")
    )

    server <- function(input, output, session) {

    }

    shinyApp(ui, server)
}